requirements.txt
fastapi==0.115.5
python-telegram-bot==21.7
